Выполнил ДЗ по 5 модулю "Модуляция"
