+++ js/util.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+export function getRandomNumber(min, max) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
